add like/dislike/delete-card

+++ src/utils/Api.js
@@ -65,23 +65,41 @@ class Api {
     })
     .then(this._checkResponse);
   }
-
-  //поставить лайк
-  putLike(cardId) {
-    return fetch(`${this._baseUrl}/cards/${cardId}/likes`, {
-      method: 'PUT',
-      headers: this._headers,
-    })
-    .then(this._checkResponse);
+  
+  changeLikeCardStatus(cardId, likeStatus) {
+    if(likeStatus){
+      //поставить лайк
+      return fetch(`${this._baseUrl}/cards/${cardId}/likes`, {
+        method: 'PUT',
+        headers: this._headers,
+      })
+      .then(this._checkResponse);
+    } else {
+      //удалить лайк
+      return fetch(`${this._baseUrl}/cards/${cardId}/likes`, {
+        method: 'DELETE',
+        headers: this._headers,
+      })
+      .then(this._checkResponse);
+    }
   }
-  //удалить лайк
-  deleteLike(cardId) {
-    return fetch(`${this._baseUrl}/cards/${cardId}/likes`, {
-      method: 'DELETE',
-      headers: this._headers,
-    })
-    .then(this._checkResponse);
-  }
+  // //поставить лайк
+  // putLike(cardId) {
+  //   return fetch(`${this._baseUrl}/cards/${cardId}/likes`, {
+  //     method: 'PUT',
+  //     headers: this._headers,
+  //   })
+  //   .then(this._checkResponse);
+  // }
+  // //удалить лайк
+  // deleteLike(cardId) {
+  //   return fetch(`${this._baseUrl}/cards/${cardId}/likes`, {
+  //     method: 'DELETE',
+  //     headers: this._headers,
+  //   })
+  //   .then(this._checkResponse);
+  // }
+ 
   //обновление аватара пользователя
   editAvatar(data) {
     return fetch(`${this._baseUrl}/users/me/avatar`, {
